Move Resources extensions from tooling to utils
