+++ detekt-utils/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("module")
+}
